+++ implementation/colors_balance_locally_painter.cpp
@@ -31,12 +31,12 @@ void ColorsBalanceLocallyPainter::paintEdge(BAGraph& graph, BAEdge& edge) {
     ++colors_histogram_[color];
   }
   else {
-    std::uniform_real_distribution<float> distribution{0.0F, 1.0F};
-    const float r = distribution(generator_);
-    float cumulative_probability = 0.0F;
+    std::uniform_real_distribution<double> distribution{0.0F, 1.0F};
+    const double r = distribution(generator_);
+    double cumulative_probability = 0.0F;
     for (const auto& color : edges_colors_) {
-      const float probability =
-          static_cast<float>(total - local_colors_histogram[color]) / static_cast<float>((edges_colors_.size() - 1) * total);
+      const double probability =
+          static_cast<double>(total - local_colors_histogram[color]) / static_cast<double>((edges_colors_.size() - 1) * total);
       cumulative_probability += probability;
 
       if (r < cumulative_probability) {
